backend/services/api-server/pyproject.toml
[tool.poetry]
name = "API Server"
version = "0.1.0"
description = ""
authors = ["Ihar Suvorau <ihar.suvorau@gmail.com>"]
readme = "README.md"
packages = [{ include = "api_server", from = "." }]

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^2.0.20"
fastapi = "^0.103.1"
fastapi-users = { version = "^12.1.2", extras = ["sqlalchemy"] }
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
uvicorn = "^0.23.2"
httpx = "^0.25.0"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
python-multipart = "^0.0.6"
greenlet = "^2.0.2"
opentelemetry-distro = "^0.43b0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-httpx = "^0.43b0"
opentelemetry-instrumentation-requests = "^0.43b0"
opentelemetry-instrumentation-logging = "^0.43b0"
opentelemetry-exporter-otlp = "^1.21.0"
opentelemetry-exporter-prometheus = "^0.43b0"
kafka-python = "^2.0.2"
python-dotenv = "^1.0.0"
fastapi-users-db-sqlalchemy = "^6.0.1"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
poetry-plugin-export = "^1.6.0"

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
